+++ src/parser/core/modules/ResourceGraphs/index.ts
@@ -1,0 +1,2 @@
+export {ResourceGraphs} from './ResourceGraphs'
+export type {Resource, ResourceDatum} from './ResourceGraphs'
